feature: Progressive Rendering

+++ src/algebra/vector/dot.c
@@ -6,13 +6,13 @@
 /*   By: nmartins <nmartins@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/26 19:32:36 by nmartins       #+#    #+#                */
-/*   Updated: 2019/10/26 19:37:19 by nmartins      ########   odam.nl         */
+/*   Updated: 2019/11/18 10:17:08 by nmartins      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vector.h"
 
-double	vec_dot(const t_vec *a, const t_vec *b)
+REAL	vec_dot(const t_vec *a, const t_vec *b)
 {
 	return (a->x * b->x + a->y * b->y + a->z * b->z);
 }
